requirements.txt
# Core
fastapi[all]>=0.115.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Auth
python-jose[cryptography]>=3.3.0

# Config
python-dotenv>=1.0.0
pydantic-settings>=2.6.0

# Testing
httpx>=0.25.0
pytest>=7.4.0

# Dev tools
black>=23.11.0
ruff>=0.1.6
